Collection Log Luck - Separate distributions and drops into packages
