+++ src/main/java/com/evansloan/collectionlog/luck/drop/ConfiguredBinomialDrop.java
@@ -1,8 +1,9 @@
-package com.evansloan.collectionlog.luck.probability;
+package com.evansloan.collectionlog.luck.drop;
 
 import com.evansloan.collectionlog.CollectionLog;
 import com.evansloan.collectionlog.CollectionLogConfig;
 import com.evansloan.collectionlog.CollectionLogItem;
+import com.evansloan.collectionlog.luck.RollInfo;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
