+++ src/main/java/com/evansloan/collectionlog/luck/drop/MissingKillCountDrop.java
@@ -1,4 +1,4 @@
-package com.evansloan.collectionlog.luck.probability;
+package com.evansloan.collectionlog.luck.drop;
 
 // Items whose KC is not tracked by the collection log cannot currently have their luck calculated. In the future,
 // it may be possible to use the loot tracker plugin to implement some of these items.
