+++ src/main/java/com/evansloan/collectionlog/luck/drop/PoissonBinomialStackDrop.java
@@ -1,4 +1,4 @@
-package com.evansloan.collectionlog.luck.probability;
+package com.evansloan.collectionlog.luck.drop;
 
 // Some complicated distributions can be approximated by a normal distribution. For example, measuring the amount of
 // raw GP from an activity where the drop itself has multiple drop chances plus a quantity ranges.
