+++ src/test/java/com/evansloan/collectionlog/luck/drop/HiddenShardDropTest.java
@@ -1,8 +1,9 @@
-package com.evansloan.collectionlog.luck.probability;
+package com.evansloan.collectionlog.luck.drop;
 
 import com.evansloan.collectionlog.CollectionLog;
 import com.evansloan.collectionlog.CollectionLogItem;
 import com.evansloan.collectionlog.luck.LogItemSourceInfo;
+import com.evansloan.collectionlog.luck.RollInfo;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
